frontend/web/requirements.txt
dash==3.2.0
python-dotenv==1.1.1
Requests==2.32.5
